% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_time_rhythms.R
\name{compare_time_rhythms}
\alias{compare_time_rhythms}
\title{Plot and compare EEG brain rhythms from two time series}
\usage{
compare_time_rhythms(
  x = NULL,
  y = NULL,
  target_rhythm = "alpha",
  fs = 100,
  N = 600,
  flim = c(0, 0.5 * fs),
  channel_label1 = "p",
  channel_label2 = "q",
  spectrum_smoothness = 0.4
)
}
\arguments{
\item{spectrum_smoothness}{Cut-frequency (0-1) of a 3rd Butterworth LPF}
}
\description{
Plot and compare EEG brain rhythms from two time series
}
\examples{
\dontrun{
 compare_time_rhythms()
}
}
\keyword{EEG}
