% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_smooth_spectrum.R
\name{estimate_smooth_spectrum}
\alias{estimate_smooth_spectrum}
\title{Provides a smooth estimator of the spectrum of a signal}
\usage{
estimate_smooth_spectrum(x, fcut = 0.05)
}
\description{
Provides a smooth estimator of the spectrum of a signal
}
